Add initial diagram test
This adds the a inital diagram as a Draw.io file. You can use
the free open source online tool to edit the diagrams. Please
let me know if you have any questions.

In this first test, database, server storage, and app data are
given. Database refers to the sql server, while server storage
refers to the local storage on the computer. App data is data
available on the device.